requirements.txt
Bootstrap_Flask==2.2.0
Flask_CKEditor==0.4.6
WTForms~=3.0.1
Flask_WTF==1.2.1
Werkzeug==3.0.6
Flask~=2.3.2
flask_sqlalchemy==3.1.1
SQLAlchemy~=2.0.25
email_validator
python-dotenv~=1.1.0
Flask-Login~=0.6.3